update build workflow
we had a check for build but that was triggering when we merged a PR on
main branch and I don't think that's useful. We still had failures even
after that check.

This commits adds the same check to pr-lint workflow and removes the
build workflow that runs when a PR gets merged into main branch.

The reason to run it as part of lint workflow was because we have go
installed as part of this workflow.

We now also run it before we lint the codebase, this makes sure that we
run the build first followed by linting the codebase.

Signed-off-by: kranurag7 <anurag.kumar@syself.com>